tests/empty/output/requirements.txt
docutils>=0.12
django>=1.7
django-select-multiple-field>=0.3.1
Pillow>=2.7.0
djangorestframework>=2.4.4
